Note: Set selection at end. Consolidate setSelection logic.
