+++ src/components/Note.js
@@ -12,6 +12,7 @@ import setAttribute from '../action-creators/setAttribute'
 import {
   attribute,
   isContextViewActive,
+  setSelection,
 } from '../util.js'
 
 // gets the editable node for the given note element
@@ -37,16 +38,16 @@ export const Note = ({ context }) => {
         const note = attribute(context, '=note')
 
         // select thought
-        if (e.key === 'Escape' || (e.metaKey && e.altKey && e.keyCode === 'N'.charCodeAt(0))) {
+        if (e.key === 'Escape' || e.key === 'ArrowUp' || (e.metaKey && e.altKey && e.keyCode === 'N'.charCodeAt(0))) {
           e.stopPropagation()
-          editableOfNote(e.target).focus()
+          setSelection(editableOfNote(e.target), { end: true })
         }
         // delete note
         // note may be '' or null if the attribute child was deleted
         else if (e.key === 'Backspace' && (!note || (e.shiftKey && (e.metaKey || e.ctrlKey)))) {
           e.stopPropagation() // prevent delete thought
           e.preventDefault()
-          editableOfNote(e.target).focus()
+          setSelection(editableOfNote(e.target), { end: true })
           dispatch(deleteAttribute(context, '=note'))
         }
       }}
